refactor(count-hll): share hll.rs's classic estimator instead of a copy
`estimate_bucket` was a hand-copied HyperLogLog estimator. The copy had
drifted: it still carried the 32-bit large-range correction
(`est > 143_165_576.533`, folding through `-2^32 * ln(1 - est/2^32)`)
that hll.rs dropped in #128 because ranks come from a 64-bit hash and the
correction only ever distorted large estimates.

hll.rs now exposes the estimator body as `classic_estimate(&[u8])` and
`HyperLogLogImpl<Classic, _, _>::estimate` delegates to it, so the two
callers cannot drift again. The HyperLogLog numbers are unchanged:
`alpha_m * m * m * indicator()` and `alpha_m * m * m / z` are the same
expression, and the small-range branch is carried over verbatim.
